% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filter.R
\name{add_filter}
\alias{add_filter}
\title{Apply a filter on each group}
\usage{
add_filter(aggregated_data, ...)
}
\arguments{
\item{aggregated_data}{aggregated_data}

\item{...}{Filter criterion, e.g. CV < 0.1}
}
\value{
An updated aggregated_data data frame
}
\description{
This function adds a new boolean column to aggregated_data based whether a
given criterion is met (TRUE) or not (FALSE).
}
