% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data}
\usage{
aggregate_data(metabolites, meta_data, conc_values, quant_status, status_vec)
}
\arguments{
\item{metabolites}{metabolites MetAlyzer object}

\item{conc_values}{conc_values of a MetAlyzer object}

\item{quant_status}{quant_status of a MetAlyzer object}

\item{meta_columns}{A selection of columns from meta_data to add to
aggregated data frame}
}
\description{
This function reshapes conc_values, quant_status,
metatabolitesand sample IDs and combines them into a tibble data frame
for filtering with dplyr and plotting with 'ggplot2'. "aggregated_data"
is grouped by metabolites.
}
\keyword{internal}
