% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{aggregated_data}
\alias{aggregated_data}
\title{Get Aggregated Data}
\usage{
aggregated_data(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}
}
\description{
This function returns the tibble "aggregated_data".
}
\examples{
metalyzer_se <- MetAlyzer::read_metidq(conc_file_path = MetAlyzer::load_demodata_biocrates())

MetAlyzer::aggregated_data(metalyzer_se)
}
