% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregate data}
\usage{
aggregateData(
  metalyzer,
  ...,
  status_list = list(Valid = c("Valid", "LOQ"), `Valid threshold` = 0.5, KO = c())
)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to aggregated data
frame that will be used as extra grouping variables.}

\item{status_list}{A list with three entries.
"Valid" holds the vector with a quantification status that is valid.
"Valid threshold" holds the minimal percentage threshold.
"KO" holds a character vector with quantification status that directly
invalidates a set of repetitions once it is encountered.}
}
\value{
The aggregated data tibble data frame
}
\description{
This function reshapes concentration, quant_status
and meta_data and combines them in a tibble data frame.
"aggregated_data" is grouped by metabolites as well as the selection
of additional variables from meta_data. The format is suitable for
plotting with ggplot2.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())
renameMetaData(metalyzer, Method = `Sample Description`)
filterMetaData(metalyzer, !is.na(Tissue))

aggregated_data <- aggregateData(metalyzer, Tissue, Method)
}
