% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{aggregatedData}
\alias{aggregatedData}
\title{Get Aggregated Data}
\usage{
aggregatedData(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}
}
\description{
This function returns the tibble "aggregated_data".
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_extraction_data())

aggregatedData(metalyzer_se)
}
