% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_log2FC.R
\name{apply_linear_model}
\alias{apply_linear_model}
\title{Calculate log2 fold change}
\usage{
apply_linear_model(df, ...)
}
\arguments{
\item{df}{A subset data frame}

\item{...}{}
}
\description{
This function applies a linear model to calculate the log2 fold change and
its significance
}
\keyword{internal}
