% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{assign_quant_status}
\alias{assign_quant_status}
\title{Get quantification status}
\usage{
assign_quant_status(df_BG)
}
\arguments{
\item{df_BG}{A data frame containing background color information}
}
\value{

}
\description{
This function assigns the background color of each cell in .full_sheet to the
corresponding quantification status
}
