% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{calc_anova}
\alias{calc_anova}
\title{Perform an ANOVA}
\usage{
calc_anova(c_vec, d_vec, valid_vec)
}
\arguments{
\item{c_vec}{A character vector containing the categorical variables}

\item{d_vec}{A numeric vector containing the dependent variables}

\item{valid_vec}{A logical vector for filtering}
}
\description{
This function filters based on the filter vector valid_vec and then performs
a one-way ANOVA
}
