% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{calc_CV}
\alias{calc_CV}
\title{Add CV}
\usage{
calc_CV(aggregated_data, cv_threshs)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}

\item{cv_threshs}{A numeric vector of thresholds}
}
\description{
This function adds the mean, standard deviation (SD) and the
coefficient of variation (CV) to aggregated_data
}
\keyword{internal}
