% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{calc_log2FC}
\alias{calc_log2FC}
\title{Differential analysis}
\usage{
calc_log2FC(metalyzer_se, group, group_level = NULL)
}
\arguments{
\item{metalyzer_se}{An SE object output from \code{\link[MetAlyzer]{read_metidq()}}.}

\item{group}{A character specifying the sample metadata column containing two
groups that will be compared.}

\item{group_level}{A length-2 vector of characters specifying the group members
in `group`, which decides the direction of comparisons. For example, c('A', 'B')
compares Group A to B, and vice versa. Default is NULL (alphabetically).}
}
\value{
An SE object with an added table of differential analysis results, accessible
via `metalyzer_se@metadata$log2FC`.
}
\description{
Perform differential analysis and add the results table to the input
`SummarizedExperiment` (SE) object. The analysis is conducted using the \pkg{limma}
package, yielding log2 fold changes, p-values, and adjusted p-values. Note that
the input data must already be log2 transformed, and should be subset if the
variable of interest contains more than two groups.
}
\keyword{internal}
