% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_calculate_log2FC.R
\name{calculate_log2FC}
\alias{calculate_log2FC}
\title{Calculate log2 fold change}
\usage{
calculate_log2FC(aggregated_data, categorical, installation_type = "binary")
}
\arguments{
\item{aggregated_data}{aggregated_data}

\item{categorical}{A column specifying the two groups}

\item{installation_type}{A character, indicating the type of package to
download and install. Options: ["binary", "source", "both"]}
}
\value{
A data frame containing the log2 fold change for each metabolite
}
\description{
This function calculates the log2 fold change of two groups from
plotting_data.
}
\examples{
\dontrun{
print(1)
}
}
