% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plotting_data.R
\name{create_plotting_data}
\alias{create_plotting_data}
\title{Create plotting data}
\usage{
create_plotting_data(object, ..., ungrouped, ts, valid_vec, t)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to reshaped data frame}

\item{ungrouped}{A column from meta_data to add to reshaped data frame that
will not be used as grouping variables}

\item{ts}{A numeric vector of thresholds for CV categorization}

\item{valid_vec}{A character vector containing each quantification status that
is considered to be a valid measurement}

\item{t}{A numeric threshold to determine valid measurements}
}
\description{
This function reshapes raw_data, quant_status and meta_data and combines them
together with basic statistics in a tibble data frame for plotting with
'ggplot2'. plotting_data is grouped by metabolites as well as the selection of
additional variables. Statistics are then calculated per group.
}
\keyword{internal}
