% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{create_viridis_style}
\alias{create_viridis_style}
\title{Creates a viridis color style for Plotly plots.}
\usage{
create_viridis_style(
  color_scale,
  type = "scale",
  data = NULL,
  values_col_name = NULL
)
}
\arguments{
\item{color_scale}{The name of the palette (e.g., "Magma", "Viridis").}

\item{type}{The desired output type: "scale" (for a color bar), "hex" or "inital"
(for a color scalde, a vector of hex codes, or the correct scale for viridis package). Defaults to "scale".}

\item{data}{The data frame containing the values. Only required if type = "hex".}

\item{values_col_name}{The name of the column with numeric values. Only
required if type = "hex".}
}
\value{
A data frame if type is "scale", or a character vector if type is "hex".
}
\description{
This function can generate either a Plotly-compatible colorscale for a
color bar or a vector of hex color codes for manual coloring.
}
