% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{createPlottingData}
\alias{createPlottingData}
\title{Create plotting data}
\usage{
createPlottingData(
  MetAlyzer,
  ...,
  ts = c(0.1, 0.2, 0.3),
  valid_vec = c("Valid", "LOQ"),
  t = 0.5
)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to reshaped data frame}

\item{ts}{A numeric vector of thresholds for CV categorization}

\item{valid_vec}{A character vector containing each quantification status that
is considered to be a valid measurement}

\item{t}{A numeric threshold to determine valid measurements}
}
\value{
An updated MetAlyzer object
}
\description{
This method combines raw_data, quant_status and meta_data to one tibble
data frame
}
\examples{
\dontrun{
obj <- createPlottingData(obj, Tissue, Method,
ts = c(0.1, 0.2, 0.3),
valid_vec = c("Valid", "LOQ"), t = 0.5)
}
}
