% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imputation.R
\name{data_imputation}
\alias{data_imputation}
\title{Impute aggregated data}
\usage{
data_imputation(metalyzer_se, impute_perc_of_min, impute_NA)
}
\arguments{
\item{metalyzer_se}{A MetAlyzer object}

\item{impute_perc_of_min}{A numeric value between 0 and 1. Percentage
of the minimal positive value, that is taken for imputation}

\item{impute_NA}{Logical value whether to impute NA values}
}
\description{
This function imputes zero concentration values
with a percentage of the minimal positive value. If all values
are zero or NA, they are set to NA. The imputed values are added
to plotting_data in an extra column imputed_Conc.
}
\keyword{internal}
