% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{data_normalization}
\alias{data_normalization}
\title{Normalization}
\usage{
data_normalization(metalyzer_se, norm_method = "log2")
}
\arguments{
\item{metalyzer_se}{An SE object output from \code{\link{read_webidq}}.}

\item{norm_method}{A character specifying the normalization method to use, which
should be one of 'log2' (default), 'median', or 'TIC'.}
}
\value{
An SE object with the normalized `Concentration` column in the aggregated
data accessible via `metalyzer_se@metadata$aggregated_data`.
}
\description{
Normalize concentration values among samples using glog2 transformation,
median normalization, or total ion count (TIC) normalization and update the `Concentration`
column in the aggregated data in the input `SummarizedExperiment` (SE) object.
}
\keyword{2,}
\keyword{:}
\keyword{=}
\keyword{In}
\keyword{NaNs}
\keyword{TRUE))}
\keyword{Warning}
\keyword{internal}
\keyword{log(apply(x,}
\keyword{median,}
\keyword{message:}
\keyword{na.rm}
\keyword{produced}
