% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{data_transformation}
\alias{data_transformation}
\title{Transform aggregated data}
\usage{
data_transformation(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{a MetAlyzer object}
}
\value{
An updated aggregated_data tibble data frame
}
\description{
This function performs a transformation of imputed concentration
values (imputed_Conc) with a given function. NA values are skipped. The
transformed values are added to aggregated_data in an extra column
transf_Conc.
}
\keyword{internal}
