% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way_anova.R
\name{do_anova_2w}
\alias{do_anova_2w}
\title{Two way anova function}
\usage{
do_anova_2w(
  aggregate_obj,
  categorical,
  log2_transformation = "TRUE",
  impute = "FALSE",
  first_con = "drosophila",
  second_con = "mouse"
)
}
\description{
This function allows you to perform two-way anova.
}
