% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_FC_network.R
\name{draw_FC_network}
\alias{draw_FC_network}
\title{Draw metabolic pathway}
\usage{
draw_FC_network(
  log2FC_df,
  output,
  q_value = 0.05,
  figsize = c(40, 36),
  dpi = 300,
  bg_color = "#FFFFFF",
  na_color = "#D3DDDC",
  node_size = 500,
  font_size = 15,
  font_color = "#000000",
  edge_color = "#757575",
  ann_font_size = 25,
  ann_font_color = "#046C9A",
  font_family = "sans-serif",
  pathway_file = system.file("extdata", "pathway.xlsx", package = "MetAlyzer"),
  path_to_conda = NULL
)
}
\arguments{
\item{log2FC_df}{log2FC_df}

\item{output}{output}

\item{q_value}{q_value}

\item{figsize}{figsize}

\item{bg_color}{bg_color}

\item{node_size}{node_size}

\item{font_size}{font_size}

\item{font_color}{font_color}

\item{edge_color}{edge_color}

\item{ann_font_size}{ann_font_size}

\item{ann_font_color}{ann_font_color}

\item{pathway_file}{pathway_file}

\item{path_to_conda}{path_to_conda}

\item{colbar_width}{colbar_width}

\item{colors}{colors}
}
\description{
This function ...
}
\examples{
\dontrun{
print(1)
}
}
