% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{dropInvalidMetabolites}
\alias{dropInvalidMetabolites}
\title{Drop invalid metabolites}
\usage{
dropInvalidMetabolites(metalyzer, valid_perc = 0)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{valid_perc}{A numeric lower threshold between 0 and 1 (t < x) to
determine valid metabolites with a given percentage of valid replicates.}
}
\description{
This function extracts metabolites with a given percentage
of valid replicates from aggregated_data and removes them from
metabolites and aggregated data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())
renameMetaData(metalyzer, Method = `Sample Description`)
filterMetaData(metalyzer, !is.na(Tissue))
aggregateData(metalyzer, Tissue, Method)

dropInvalidMetabolites(metalyzer)
}
