% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{example_extraction_data}
\alias{example_extraction_data}
\title{Get example extraction data}
\usage{
example_extraction_data()
}
\value{
extraction_data_MxP_Quant_500.xlsx file path
}
\description{
This function returns the extraction_data_MxP_Quant_500.xlsx file path.
}
\examples{
fpath <- example_extraction_data()
}
