% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{example_meta_data}
\alias{example_meta_data}
\title{Get example meta data}
\usage{
example_meta_data()
}
\value{
data frame loaded from example_meta_data.RDS
}
\description{
This function returns the data frame loaded from example_meta_data.RDS.
}
\examples{
fpath <- example_meta_data()
}
