% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{example_mutation_data_xl}
\alias{example_mutation_data_xl}
\title{Get example mutation data}
\usage{
example_mutation_data_xl()
}
\value{
mutation_data_MxP_Quant_500_XL.xlsx file path
}
\description{
This function returns the mutation_data_MxP_Quant_500_XL.xlsx file path.
}
\examples{
fpath <- example_mutation_data_xl()
}
