% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{example_treatment_data}
\alias{example_treatment_data}
\title{Get example treatment data}
\usage{
example_treatment_data()
}
\value{
treatment_data_MxP_Quant_500.xlsx file path
}
\description{
This function returns the treatment_data_MxP_Quant_500.xlsx file path.
}
\examples{
fpath <- example_treatment_data()
}
