% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_aggregated_data.R
\name{expand_aggregated_data}
\alias{expand_aggregated_data}
\title{Add column to aggregated_data}
\usage{
expand_aggregated_data(metalyzer_se, meta_data_column)
}
\arguments{
\item{metalyzer_se}{A Metalyzer object}

\item{meta_data_column}{A column from meta data}
}
\value{
An updated Metalyzer object
}
\description{
This function adds a column to the aggregated_data tibble by
using information from the meta data.
}
\keyword{internal}
