% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_handler.R
\name{export_conc_values}
\alias{export_conc_values}
\title{Export filtered raw data as csv}
\usage{
export_conc_values(metalyzer_se, ..., file_path = "metabolomics_data.csv")
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Additional columns from meta_data}

\item{file_path}{file path}
}
\description{
This function exports the filtered raw data in the CSV format.
}
\examples{
metalyzer_se <- read_metidq(file_path = load_rawdata_extraction())

output_file <- file.path(tempdir(), "metabolomics_data.csv")
export_conc_values(
  metalyzer_se,
  `Sample Description`,
  Tissue,
  file_path = output_file
)
unlink(output_file)
}
