% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{export_raw_data}
\alias{export_raw_data}
\title{Export filtered raw data as csv}
\usage{
export_raw_data(object, sample_id, file_path)
}
\arguments{
\item{object}{MetAlyzer object}

\item{sample_id}{A column from meta_data for sample identification}

\item{file_path}{file path}
}
\description{
This function exports the filtered raw data in the CSV format.
}
\keyword{internal}
