% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{exportConcValues}
\alias{exportConcValues}
\title{Export filtered raw data as csv}
\usage{
exportConcValues(metalyzer, ..., file_path = "metabolomics_data.csv")
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{...}{Additional columns from meta_data}

\item{file_path}{file path}
}
\description{
This function exports the filtered raw data in the CSV format.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

exportConcValues(metalyzer, `Sample Description`, Tissue)
}
