% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_exportRawData-method.R
\name{exportRawData}
\alias{exportRawData}
\alias{exportRawData,MetAlyzer-method}
\title{Export filtered raw data as csv}
\usage{
exportRawData(object, sample_id, file_path)

\S4method{exportRawData}{MetAlyzer}(object, sample_id, file_path)
}
\arguments{
\item{object}{MetAlyzer object}

\item{sample_id}{A column from meta_data for sample identification}

\item{file_path}{file path}
}
\description{
This method exports the filtered raw data in the CSV format.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Export filtered raw data as csv
}}

\examples{
\dontrun{
print(1)
}
}
