% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{extraction_data}
\alias{extraction_data}
\title{Get example data file path}
\usage{
extraction_data()
}
\value{
extraction_data.xlsx file path
}
\description{
This function returns the extraction_data.xlsx file path.
}
\examples{
fpath <- extraction_data()
}
