% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_classes}
\alias{filter_classes}
\title{Filter metabolites}
\usage{
filter_classes(object, class_name = "Metabolism Indicators")
}
\arguments{
\item{object}{MetAlyzer object}

\item{class_name}{A class to be filtered out}
}
\value{

}
\description{
This function filters out certain classes of the metabolites vector and raw_data and quant_status columns
}
