% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_groupwise.R
\name{filter_groupwise}
\alias{filter_groupwise}
\title{Apply a filter on each group}
\usage{
filter_groupwise(
  aggregated_data,
  ...,
  filter_col = .data$Valid_Replicates,
  lower_bound = 0,
  upper_bound = FALSE
)
}
\arguments{
\item{aggregated_data}{aggregated_data}

\item{...}{Columns of aggregated_data to group by}

\item{filter_col}{Column with logical values that defines whether a filter is
met or not}

\item{lower_bound}{Numerical threshold between 0 and 1 defining the lower
bound (t < x)}

\item{upper_bound}{Numerical threshold between 0 and 1 defining the upper
bound (x < t)}
}
\value{
A filtered aggregated_data data frame
}
\description{
This function takes a boolean column from aggregated_data and filters based
on the percentage of samples within each group that meet the filter criterion
(=TRUE).
}
\examples{
\dontrun{
print(1)
}
}
