% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_meta_data}
\alias{filter_meta_data}
\title{Filter meta data}
\usage{
filter_meta_data(object, column, keep = NULL, remove = NULL)
}
\arguments{
\item{object}{MetAlyzer object}

\item{column}{A length-one character vector specifying which column of
meta_data to use for filtering}

\item{keep}{A character vector specifying which samples to keep}

\item{remove}{A character vector specifying which samples to remove}
}
\value{

}
\description{
This function updates the "Filter" column in meta_data to filter out samples
}
