% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetaData}
\alias{filterMetaData}
\title{Filter meta data}
\usage{
filterMetaData(MetAlyzer, column, keep = NULL, remove = NULL)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{

}
\description{
This method calls filter_meta_data() and filters meta_data
}
