% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{get_color_option}
\alias{get_color_option}
\title{Get the color option based on the color scale}
\usage{
get_color_option(color_scale)
}
\arguments{
\item{color_scale}{A character string representing the name of the color scale.
It should be one of: "magma", "inferno", "plasma", "viridis", 
"cividis", "rocket", "mako", or "turbo".}
}
\value{
A character string representing the color option corresponding to the 
        provided color scale. If the color scale is not recognized, it returns `NA`.
}
\description{
This function maps a color scale name to a corresponding letter option.
The function returns a letter representing the color scale from a predefined mapping.
If the color scale provided is not recognized, the function fallsback to the viridis scale.
}
