% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{get_filtered_data}
\alias{get_filtered_data}
\title{Get filtered data}
\usage{
get_filtered_data(object, slot, verbose = TRUE)
}
\arguments{
\item{object}{MetAlyzer object}

\item{slot}{A length-one character vector specifying which data frame to slice}

\item{verbose}{If TRUE prints which filtered data frame is returned}
}
\value{

}
\description{
This function filters meta_data, raw_data or quant_status
}
