% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invalid_metabolites.R
\name{get_invalid_metabolites}
\alias{get_invalid_metabolites}
\title{Get invalid metabolites}
\usage{
get_invalid_metabolites(aggregated_data, filter_col, replicate_t)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}

\item{filter_col}{Boolean column defining whether the filter criterion is met or not}

\item{replicate_t}{A numeric threshold}
}
\description{
This function extracts metabolites with a given percentage of valid replicates
from aggregated_data and filters metabolites.
}
\keyword{internal}
