% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{getAggregatedData}
\alias{getAggregatedData}
\title{Get aggregated data}
\usage{
getAggregatedData(metalyzer)
}
\arguments{
\item{metalyzer}{A MetAlyzer object.}
}
\value{
The aggregated data.
}
\description{
This function returns the aggregated data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())
renameMetaData(metalyzer, Method = `Sample Description`)
filterMetaData(metalyzer, !is.na(Tissue))
aggregateData(metalyzer, Tissue, Method)

aggregated_data <- getAggregatedData(metalyzer)
}
