% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{getConcValues}
\alias{getConcValues}
\title{Get concentration values}
\usage{
getConcValues(metalyzer)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
The concentration data frame
}
\description{
This function returns the concentration data frame
(samples * metabolites) with filtered samples and metabolites.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

conc_values <- getConcValues(metalyzer)
head(conc_values, c(5, 5))
}
