% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{getMetabolites}
\alias{getMetabolites}
\title{Get metabolites}
\usage{
getMetabolites(metalyzer)
}
\arguments{
\item{metalyzer}{A MetAlyzer object.}
}
\value{
The filtered metabolites vector.
}
\description{
This function returns the filtered metabolites vector.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

metabolites <- getMetabolites(metalyzer)
metabolites[1:10]
}
