% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{getMetaData}
\alias{getMetaData}
\title{Get meta data}
\usage{
getMetaData(metalyzer)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
The filtered meta_data data frame
}
\description{
This function returns the meta_data data frame
(samples * meta variables) with filtered samples and all columns.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

meta_data <- getMetaData(metalyzer)
head(meta_data)
}
