% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{getQuantStatus}
\alias{getQuantStatus}
\title{Get quantification status}
\usage{
getQuantStatus(metalyzer)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
The quant_status data frame
}
\description{
This function returns the quant_status data frame (samples *
  metabolites) with filtered samples and metabolites.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

quant_status <- getQuantStatus(metalyzer)
head(quant_status, c(5, 5))
}
