% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{glog2}
\alias{glog2}
\title{glog2 transformation}
\usage{
glog2(x)
}
\arguments{
\item{x}{A matrix or vector containing numerical values.}
}
\value{
A glog2-transformed matrix or vector.
}
\description{
Perform generalized log2 transformation on data
}
\keyword{internal}
