% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_data}
\alias{impute_data}
\title{Impute aggregated data}
\usage{
impute_data(metalyzer_se, perc_of_min, impute_NA)
}
\arguments{
\item{perc_of_min}{A numeric value between 0 and 1. Percentage
of the minimal positive value, that is taken for imputation}

\item{impute_NA}{Logical value whether to impute NA values}

\item{metalyzer}{A MetAlyzer object}
}
\value{
An updated aggregated_data tibble data frame
}
\description{
This function imputes zero concentration values
with a percentage of the minimal positive value. If all values
are zero or NA, they are set to NA. The imputed values are added
to plotting_data in an extra column imputed_Conc.
}
