% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_plotting_data.R
\name{impute_plotting_data}
\alias{impute_plotting_data}
\title{Impute plotting data}
\usage{
impute_plotting_data(object, i, imputeNA, ...)
}
\arguments{
\item{object}{MetAlyzer object}

\item{i}{A numeric value below 1}

\item{imputeNA}{Logical value whether to impute NA values}

\item{...}{Variables to group by}
}
\description{
This function imputes zero concentration values (Concentration) with the
minimal positive value multiplied by i. If all values are zero or NA, they
are set to NA. The imputed values are added to plotting_data in an extra
column imp_Conc.
}
\keyword{internal}
