% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{init}
\alias{init}
\title{Initialize}
\usage{
init(MetAlyzer, file_path, sheet = 1)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{file_path}{A length-one character vector with the file path}

\item{sheet}{A length-one numeric vector with the sheet number}
}
\value{
An updated MetAlyzer object
}
\description{
This method initializes file path and sheet number
}
\examples{
\dontrun{
obj <- init(obj, file_path = "toydata.xlsx", sheet = 1)
}
}
