% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{initialize}
\alias{initialize}
\title{Initialize}
\usage{
initialize(object, file_path, sheet = 1)
}
\arguments{
\item{object}{MetAlyzer object}

\item{file_path}{A length-one character vector with the file path}

\item{sheet}{A length-one numeric vector with the sheet number}
}
\value{

}
\description{
This function initializes file path and sheet number
}
