% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_fpaths.R
\name{load_demodata_biocrates}
\alias{load_demodata_biocrates}
\title{Get demodata from biocrates}
\usage{
load_demodata_biocrates()
}
\value{
Metalyzer_demo dataset_biocrates MxP Quant 500 XL_2025-04.xlsx file path
}
\description{
This function returns the Metalyzer_demo dataset_biocrates MxP Quant 500 XL_2025-04.xlsx file path.
}
\examples{
fpath <- load_demodata_biocrates()
}
