% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_fpaths.R
\name{load_diffres}
\alias{load_diffres}
\title{Get example log2fc data}
\usage{
load_diffres()
}
\value{
load_diffres.rds file path
}
\description{
This function returns the log2fc dataframe of the mutation_data_MxP_Quant_500_XL.xlsx file, created with the MetaVizPro package.
}
\examples{
fpath <- load_diffres()
}
