% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{load_rawdata_extraction}
\alias{load_rawdata_extraction}
\title{Get example extraction data}
\usage{
load_rawdata_extraction()
}
\value{
extraction_data_MxP_Quant_500.xlsx file path
}
\description{
This function returns the extraction_data_MxP_Quant_500.xlsx file path.
}
\examples{
fpath <- load_rawdata_extraction()
}
