% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{log_transform}
\alias{log_transform}
\title{Logarithmic transformation}
\usage{
log_transform(vec, func)
}
\arguments{
\item{vec}{MetAlyzer object}

\item{func}{A logarithmic function}
}
\description{
This function performs logarithmic transformation of imputed concentration
values (imp_Conc)
}
\keyword{internal}
