% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_MetAlyzer-class.R
\docType{class}
\name{MetAlyzer-class}
\alias{MetAlyzer-class}
\alias{MetAlyzer}
\title{A S4 class to read and analyze 'MetIDQ' output}
\description{
A S4 class to read and analyze 'MetIDQ' output
}
\section{Slots}{

\describe{
\item{\code{file_path}}{A length-one character vector giving the file path}

\item{\code{sheet}}{A length-one numeric vector giving the sheet index}

\item{\code{metabolites}}{A character vector with all 630 measured metabolites and
optional 234 additional metabolism indicators}

\item{\code{meta_data}}{A data frame containing any meta data;
dimension: # samples x # meta variables}

\item{\code{raw_data}}{A data frame containing all raw measurements;
dimension: # samples x # metabolites}

\item{\code{quant_status}}{A data frame containing the quantification status of all
measurements; dimension: # samples x # metabolites}
}}

\examples{
\dontrun{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
data <- MetAlyzer_dataset(file_path = fpath)
}
}
