% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\docType{class}
\name{MetAlyzer}
\alias{MetAlyzer}
\title{MetAlyzer Class}
\description{
A S4 class to read and analyze 'MetIDQ' output
}
\section{Slots}{

\describe{
\item{\code{file_path}}{A length-one character vector giving the file path.}

\item{\code{sheet}}{A length-one numeric vector giving the sheet index.}

\item{\code{metabolites}}{A named list of two character vectors, named "original"
and "filtered", containing all 630 measured metabolites and optional
234 additional metabolism indicators.}

\item{\code{meta_data}}{A data frame containing any meta data.
Dimension: (samples * meta variables).}

\item{\code{conc_values}}{A data frame containing measured concentration values
for each metabolite of each sample. Dimension: (samples * metabolites).}

\item{\code{quant_status}}{A data frame containing the quantification status of
each measurement. Dimension: (samples * metabolites).}

\item{\code{aggregated_data}}{A tibble data frame combining selected variables
from meta_data, metabolite names and classes, conc_values and
quant_status in separate columns.}
}}

