% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{metalyzer_colors}
\alias{metalyzer_colors}
\title{Get MetAlyzer colors}
\usage{
metalyzer_colors()
}
\value{
data frame loaded from metalyzer_colors.RDS
}
\description{
This function returns the vector loaded from metalyzer_colors.RDS.
}
\examples{
fpath <- metalyzer_colors()
}
