% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{MetAlyzer_dataset}
\alias{MetAlyzer_dataset}
\title{Open file and read data}
\usage{
MetAlyzer_dataset(file_path, sheet = 1)
}
\arguments{
\item{file_path}{file path}

\item{sheet}{sheet index}
}
\value{
An MetAlyzer object
}
\description{
This function creates a 'MetAlyzer' object, opens the given 'MetIDQ' output Excel
sheet and extracts metabolites, raw data, quantification status and meta data.
The column "Sample Type" and the row "Class" are used as anchor cells in the
Excel sheet and are therefore a requirement.
}
\examples{
\dontrun{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
data <- MetAlyzerDataset(file_path = fpath)
}
}
