% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{MetAlyzerDataset}
\alias{MetAlyzerDataset}
\title{Open file and read data}
\usage{
MetAlyzerDataset(file_path, sheet = 1)
}
\arguments{
\item{file_path}{file path}

\item{sheet}{sheet index}
}
\value{
An MetAlyzer object
}
\description{
This function creates a MetAlyzer object, opens the given MetIDQ output Excel
file and extracts metabolites, raw data, quantification status and meta data
}
\examples{
\dontrun{
obj <- MetAlyzerDataset(file_path = "toydata.xlsx")
}
}
