% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{open_file}
\alias{open_file}
\title{Open Excel file}
\usage{
open_file(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function opens the given MetIDQ output Excel file and reads the full
given sheet.
}
\keyword{internal}
