% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{pathway}
\alias{pathway}
\title{Get pathway file path}
\usage{
pathway()
}
\value{
pathway.xlsx file path
}
\description{
This function returns the pathway.xlsx file path.
}
\examples{
fpath <- pathway()
}
