% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_file_path.R
\name{pathway_file}
\alias{pathway_file}
\title{Get pathway file path}
\usage{
pathway_file()
}
\value{
pathway.xlsx file path
}
\description{
This function returns the pathway.xlsx file path.
}
\examples{
pathway_file()
}
