% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ANOVA.R
\name{perform_ANOVA}
\alias{perform_ANOVA}
\title{One-way ANOVA}
\usage{
perform_ANOVA(aggregated_data, categorical)
}
\arguments{
\item{aggregated_data}{aggregated_data}

\item{categorical}{A column defining the categorical variable}
}
\value{
A data frame containing the log2 fold change for each metabolite
}
\description{
This method performs a one-way ANOVA on the grouped aggregated_data (the
categorical variable is removed from grouping first). The vector of the
categorical variable needs to have at least two levels after removing NAs
from the dependent variable vector. Otherwise a vector of NA is returned.
A Tukey post-hoc test is then used to determine group names, starting with
"A" followed by further letters. These group names are added to
aggregated_data in the column ANOVA_Group. Thereby, metabolites can be
identified which are significantly higher in one or more of the categorical
variable compared to all other for each metabolite.
}
\examples{
\dontrun{
print(1)
}
}
