% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{performANOVA}
\alias{performANOVA}
\title{ANOVA}
\usage{
performANOVA(MetAlyzer, categorical)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{categorical}{A  column defining the categorical variable}
}
\value{
An updated MetAlyzer object
}
\description{
This method filters based on the filter $Valid, performs a one-way ANOVA and
adds the column Group to plotting_data with the results of a Tukey post-hoc
test
}
\examples{
\dontrun{
obj <- performANOVA(obj, Methods)
}
}
