% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way_anova.R
\name{plot_2w_anv}
\alias{plot_2w_anv}
\title{Plotting the results of two way anova}
\usage{
plot_2w_anv(two_ways_res, first_con = "drosophila", second_con = "mouse")
}
\description{
This function allows you to plot the results of two-way anova.
}
