% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot Pathway Network}
\usage{
plot_network(log2FC_df, q_value = 0.05)
}
\arguments{
\item{log2FC_df}{A data frame with the log2 fold change for each metabolite}

\item{q_value}{The q-value threshold for significance}
}
\value{
ggplot object
}
\description{
This function plots the log2 fold change for each metabolite and visualizes it, in a pathway network.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_extraction_data())
metalyzer_se <- renameMetaData(metalyzer_se, Method = 'Sample Description')

log2FC_df <- calculate_log2FC(metalyzer_se, Method, impute_perc_of_min = 0.2, impute_NA = TRUE)

network <- plot_network(log2FC_df, q_value = 0.05)
}
