% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log2FC.R
\name{plot_vulcano}
\alias{plot_vulcano}
\title{Vulcano Plot Visualization}
\usage{
plot_vulcano(
  log2fc_df,
  x_cutoff = 1.5,
  y_cutoff = 0.05,
  show_labels_for = NULL
)
}
\arguments{
\item{log2fc_df}{DF with metabolites as row names and columns including log2FC, Class, qval columns.}

\item{x_cutoff}{Number of the desired log fold change cutoff for assessing significance.}

\item{y_cutoff}{Number of the desired p value cutoff for assessing significance.}

\item{show_labels_for}{Vector with Strings of Metabolite names or classes.}
}
\value{
ggplot object
}
\description{
This method creates a vulcano plot of the log2 fold change for each metabolite.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_mutation_data_xl())
metalyzer_se <- filterMetabolites(
  metalyzer_se,
  drop_metabolites = "Metabolism Indicators"
)
metalyzer_se <- renameMetaData(
  metalyzer_se,
  Mutant_Control = "Sample Description"
)
metalyzer_se <- calculate_log2FC(
  metalyzer_se,
  categorical = "Mutant_Control",
  impute_perc_of_min = 0.2,
  impute_NA = TRUE
)

log2fc_df <- metalyzer_se@metadata$log2FC
vulcano <- plot_vulcano(log2fc_df)
}
