% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plotting_data}
\alias{plotting_data}
\title{Reshape data}
\usage{
plotting_data(object, ...)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to reshaped data frame}
}
\value{

}
\description{
This function reshapes raw_data, quant_status and meta_data and combines them
in a tibble data frame for plotting with ggplot2.
}
