% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{polarity}
\alias{polarity}
\title{Get polarity file path}
\usage{
polarity()
}
\value{
polarity.csv file path
}
\description{
This function returns the polarity.csv file path.
}
\examples{
fpath <- polarity()
}
