% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_quant_data.R
\name{print_number}
\alias{print_number}
\title{Print number of quantification status}
\usage{
print_number(quant_status, status, total)
}
\arguments{
\item{quant_status}{quant_status data frame}

\item{status}{quantification status}

\item{total}{total n}
}
\description{
Print number of quantification status
}
\keyword{internal}
