% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_getter-method.R
\name{rawData}
\alias{rawData}
\alias{rawData,MetAlyzer-method}
\title{Get raw data}
\usage{
rawData(object)

\S4method{rawData}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The raw_data data frame
}
\description{
This method returns the raw_data data frame with filtered samples (rows) and
metabolites (columns).
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get raw data
}}

\examples{
\dontrun{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
data <- MetAlyzerDataset(file_path = fpath)

raw_data <- rawData(data)
head(raw_data, c(5, 5))
}
}
