% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_BG_color}
\alias{read_BG_color}
\title{Get background color}
\usage{
read_BG_color(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function gets the background color of each cell in full_sheet
}
