% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_meta_data}
\alias{read_meta_data}
\title{Get meta data}
\usage{
read_meta_data(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function slices meta data from full_sheet into meta_data
}
