% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_metabolties}
\alias{read_metabolties}
\title{Get metabolites}
\usage{
read_metabolties(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function extracts metabolites with their corresponding class from
.full_sheet into metabolites
}
