% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{read_quant_status}
\alias{read_quant_status}
\title{Read quantification status}
\usage{
read_quant_status(
  file_path,
  sheet,
  n_row,
  n_col,
  data_ranges,
  metabolites,
  status_list
)
}
\arguments{
\item{file_path}{file_path}

\item{sheet}{sheet}

\item{n_row}{n_row}

\item{n_col}{n_col}

\item{data_ranges}{data_ranges}

\item{metabolites}{metabolites}
}
\description{
This function gets the background color of each cell in .full_sheet and
assigns it to the corresponding quantification status.
}
\keyword{internal}
