% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Get raw data}
\usage{
read_raw_data(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function slices measurements from .full_sheet into raw_data
}
\keyword{internal}
