% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{readData}
\alias{readData}
\title{Open file and read data}
\usage{
readData(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function opens the given MetIDQ output Excel file and extracts metabolites, raw data and meta data
}
