% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{readQuantStatus}
\alias{readQuantStatus}
\title{Read quantification status}
\usage{
readQuantStatus(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{

}
\description{
This method gets the background color of each cell of the opened Excel file
and assigns it to the corresponding quantification status
}
