% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_handler.R
\name{rename_meta_data}
\alias{rename_meta_data}
\title{Rename meta data}
\usage{
rename_meta_data(metalyzer_se, ..., inplace = FALSE)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Use new_name = old_name to rename selected variables}

\item{inplace}{If FALSE, return a copy. Otherwise, do operation inplace
and return None.}
}
\value{
An updated SummarizedExperiment
}
\description{
This function renames a column of meta_data.
}
\examples{
metalyzer_se <- read_metidq(file_path = load_rawdata_extraction())

metalyzer_se <- rename_meta_data(
  metalyzer_se,
  Method = `Sample Description`
)
}
