% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_methods.R
\name{renameMetaData}
\alias{renameMetaData}
\alias{renameMetaData,MetAlyzer-method}
\title{Rename meta data}
\usage{
renameMetaData(object, ...)

\S4method{renameMetaData}{MetAlyzer}(object, ...)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{Use new_name = old_name to rename selected variables}
}
\value{
An updated MetAlyzer object
}
\description{
This method renames a column of meta_data using rename {dplyr}.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Rename meta data
}}

\examples{
\dontrun{
obj <- renameMetaData(obj, Method = Group)
}
}
