% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{resetMetaData}
\alias{resetMetaData}
\title{Reset meta data}
\usage{
resetMetaData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
An updated MetAlyzer object
}
\description{
This method resets the filter of meta_data
}
\examples{
\dontrun{
obj <- resetMetaData(obj)
}
}
