% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_aggregate.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshape data}
\usage{
reshape_data(metabolites, meta_columns, conc_values, quant_status)
}
\arguments{
\item{metabolites}{metabolites MetAlyzer object}

\item{meta_columns}{A selection of columns from meta_data to add to
aggregated data frame}

\item{conc_values}{conc_values of a MetAlyzer object}

\item{quant_status}{quant_status of a MetAlyzer object}
}
\description{
This function reshapes selected meta_data variables,
conc_values, quant_status and metatabolites and combines them into a
tibble data frame for filtering with dplyr and plotting with 'ggplot2'.
}
\keyword{internal}
