% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{save_plot}
\alias{save_plot}
\title{Save plots}
\usage{
save_plot(
  plot,
  folder_name = format(Sys.Date(), "\%Y-\%m-\%d"),
  folder_path = NULL,
  file_name = "network",
  format = "pdf",
  units = "cm",
  height = 21,
  width = 29.7,
  overwrite = FALSE
)
}
\arguments{
\item{plot}{A ggplot object to be saved.}

\item{folder_name}{Name of the folder where the plot will be saved. Special characters will be removed automatically. \strong{Default = date}}

\item{folder_path}{\emph{Optional: } User-defined path where the folder should be created. 
If not provided, results will be saved in `MetAlyzer_results` within the working directory. \strong{Default = NULL}}

\item{file_name}{Name of the output file (without extension). \strong{Default = "network"}}

\item{format}{File format for saving the plot (e.g., "png", "pdf", "svg"). \strong{Default = "pdf"}}

\item{units}{Units for width and height (e.g., "in", "cm", "mm"). \strong{Default = "cm"}}

\item{height}{Height of the saved plot in specified units. \strong{Default = 21.0}}

\item{width}{Width of the saved plot in specified units. \strong{Default = 29.7}}

\item{overwrite}{Logical: If `TRUE`, overwrite existing files without asking. If `FALSE`, prompt user before overwriting. \strong{Default = FALSE}}
}
\value{
The function does not return anything but saves the plot to the specified directory.
}
\description{
This function saves a given ggplot object to a specified folder and file format.
It ensures that the folder structure exists and cleans the folder name to remove special characters.
}
\keyword{ggplot}
\keyword{plot,}
\keyword{save,}
