% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{setPlottingData}
\alias{setPlottingData}
\title{Update plotting data}
\usage{
setPlottingData(MetAlyzer, plotting_data)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{plotting_data}{An updated plotting_data tibble data frame}
}
\value{
An updated MetAlyzer object
}
\description{
This method replaces plotting_data with an updated version
}
\examples{
\dontrun{
obj <- setPlottingData(obj, updated_plotting_data)
}
}
