% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{show,MetAlyzer-method}
\alias{show,MetAlyzer-method}
\title{Show a MetAlyzer object}
\usage{
\S4method{show}{MetAlyzer}(object)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{

}
\description{
This method calls show_obj() and shows the MetAlyzer object
}
