% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{show_obj}
\alias{show_obj}
\title{Show a MetAlyzer object}
\usage{
show_obj(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function shows a summary of MetAlyzer slot values
}
