% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{slice_conc_values}
\alias{slice_conc_values}
\title{Slice concentration values}
\usage{
slice_conc_values(full_sheet, data_ranges, metabolites)
}
\arguments{
\item{full_sheet}{full_sheet}

\item{data_ranges}{data_ranges}

\item{metabolites}{metabolites}
}
\description{
This function slices measurements from .full_sheet into
conc_values.
}
\keyword{internal}
