% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{slice_meta_data}
\alias{slice_meta_data}
\title{Slice meta data}
\usage{
slice_meta_data(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function slices meta data from .full_sheet into meta_data.
}
\keyword{internal}
