% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{slice_metabolites}
\alias{slice_metabolites}
\title{Slice metabolites}
\usage{
slice_metabolites(full_sheet, data_ranges)
}
\arguments{
\item{full_sheet}{full_sheet}

\item{data_ranges}{data_ranges}
}
\description{
This function extracts metabolites with their corresponding
metabolite class from .full_sheet into metabolites.
}
\keyword{internal}
