% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_MetAlyzer_dataset.R
\name{slice_metabolties}
\alias{slice_metabolties}
\title{Slice metabolites}
\usage{
slice_metabolties(full_sheet, data_ranges)
}
\arguments{
\item{full_sheet}{full_sheet}

\item{data_ranges}{data_ranges}
}
\description{
This function extracts metabolites with their corresponding metabolite class
from .full_sheet into metabolites.
}
\keyword{internal}
