% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_MetAlyzer_dataset.R
\name{slice_raw_data}
\alias{slice_raw_data}
\title{Slice raw data}
\usage{
slice_raw_data(full_sheet, data_ranges, metabolites)
}
\arguments{
\item{full_sheet}{full_sheet}

\item{data_ranges}{data_ranges}

\item{metabolites}{metabolites}
}
\description{
This function slices measurements from .full_sheet into raw_data.
}
\keyword{internal}
