% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{slice_status_values}
\alias{slice_status_values}
\title{Slice quantification status values}
\usage{
slice_status_values(full_sheet, data_ranges, metabolites)
}
\arguments{
\item{full_sheet}{A data frame or matrix representing the full Excel sheet
containing the status values.}

\item{data_ranges}{A list containing the row and column indices for the data.}

\item{metabolites}{A character vector of metabolite names to be used as the
column headers of the output data frame.}
}
\description{
This function slices quant status from .full_sheet into
status_values.
}
\keyword{internal}
