% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{sum_quant_data}
\alias{sum_quant_data}
\title{Summarize quantification status}
\usage{
sum_quant_data(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function lists the number of each quantification status and its
percentage.
}
\keyword{internal}
