% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{summariseConcValues}
\alias{summariseConcValues}
\title{Summarise concentration values}
\usage{
summariseConcValues(metalyzer)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
A vector with metabolites that contain NAs
}
\description{
This function prints quantiles and NAs of raw data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

na_metabolites <- summariseConcValues(metalyzer)
na_metabolites
}
