% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{summarize_quant_data}
\alias{summarize_quant_data}
\title{Summarize quantification status}
\usage{
summarize_quant_data(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}
}
\description{
This function lists the number of each quantification status and
its percentage.
}
\examples{
metalyzer_se <- MetAlyzer::read_metidq(file_path = MetAlyzer::load_demodata_biocrates())

MetAlyzer::summarize_quant_data(metalyzer_se)
}
