% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{summarizeConcValues}
\alias{summarizeConcValues}
\title{Summarize concentration values}
\usage{
summarizeConcValues(metalyzer)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
A vector with metabolites that contain NAs
}
\description{
This function prints quantiles and NAs of raw data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

na_metabolites <- summarizeConcValues(metalyzer)
na_metabolites
}
