% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_fpaths.R
\name{toy_diffres}
\alias{toy_diffres}
\title{Get example log2fc data}
\usage{
toy_diffres()
}
\value{
toy_diffres.rds file path
}
\description{
This function returns the log2fc dataframe of the mutation_data_MxP_Quant_500_XL.xlsx file, created with the MetaVizPro package.
}
\examples{
fpath <- toy_diffres()
}
