% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform aggregated data}
\usage{
transform_data(aggregated_data)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}
}
\value{
An updated aggregated_data tibble data frame
}
\description{
This function performs a transformation of imputed concentration
values (imputed_Conc) with a given function. NA values are skipped. The
transformed values are added to aggregated_data in an extra column
transf_Conc.
}
