% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{transformPlottingData}
\alias{transformPlottingData}
\title{log transform plotting data}
\usage{
transformPlottingData(MetAlyzer, func = log2)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{func}{A logarithmic function to transform concentration values with
number of samples}
}
\value{
An updated MetAlyzer object
}
\description{
This method performs a transformation of imputed concentration values
(imp_Conc) with a given funciton. NA values are skipped. The transformed
values are added to plotting_data in an extra column transf_Conc.
}
\examples{
\dontrun{
obj <- transformPlottingData(obj, func = log2)
obj <- transformPlottingData(obj, func = log)
}
}
