% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{treatment_data}
\alias{treatment_data}
\title{Get example data file path}
\usage{
treatment_data()
}
\value{
treatment_data.xlsx file path
}
\description{
This function returns the treatment_data.xlsx file path.
}
\examples{
fpath <- treatment_data()
}
