% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{unify_hex}
\alias{unify_hex}
\title{Unify hex codes}
\usage{
unify_hex(hex)
}
\arguments{
\item{hex}{A 3, 4, 6 or 8 digit hex code}
}
\description{
This function gets hex codes of different length and returns
them in a unified format.
}
\keyword{internal}
