% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{update_meta_data}
\alias{update_meta_data}
\title{Update meta data}
\usage{
update_meta_data(object, name, new_colum)
}
\arguments{
\item{object}{MetAlyzer object}

\item{name}{The new column name}

\item{new_colum}{A vector for the new column (length has to be same as the
number of filtered samples)}
}
\description{
This function adds another column to filtered meta_data.
}
\keyword{internal}
