% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{updateMetaData}
\alias{updateMetaData}
\title{Update meta data}
\usage{
updateMetaData(MetAlyzer, name, new_colum)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{name}{A length-one character vector giving the new column name}

\item{new_colum}{A vector for the new column (length has to be same as the
number of samples)}
}
\value{
An updated MetAlyzer object
}
\description{
This method adds another column to unfiltered meta_data
}
\examples{
\dontrun{
obj <- updateMetaData(obj, "Date", format(Sys.Date()))
obj <- updateMetaData(obj, "Analyzed", TRUE)
}
}
