% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_aggregate.R
\name{valid_measurement}
\alias{valid_measurement}
\title{Add valid filter}
\usage{
valid_measurement(aggregated_data, status_list)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}

\item{status_list}{A list with three entries.
"Valid" holds the vector with a quantification status that is valid.
"Valid threshold" holds the minimal percentage threshold.
"KO" holds a character vector with quantification status that directly
invalidates a set of repetitions once it is encountered.}
}
\description{
This function adds a filter column to aggregated_data based
on the quantification status. The filter is True if the percentage of
valid measurements is at least a given threshold.
}
\keyword{internal}
