% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{xl_data}
\alias{xl_data}
\title{Get example data xl file path}
\usage{
xl_data()
}
\value{
MetIDQ_XL.xlsx file path
}
\description{
This function returns the MetIDQ_XL.xlsx file path.
}
\examples{
fpath <- xl_data()
}
