% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{zero_imputation}
\alias{zero_imputation}
\title{Zero imputation}
\usage{
zero_imputation(vec, i)
}
\arguments{
\item{vec}{A numeric vector containing the concentration values}

\item{i}{A numeric value below 1)}
}
\description{
This function performs zero imputation with the minimal positive value times i
}
\keyword{internal}
