% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plotting_data.R
\name{calc_CV}
\alias{calc_CV}
\title{Add CV}
\usage{
calc_CV(plotting_data, ts)
}
\arguments{
\item{plotting_data}{plotting_data tibble data frame}

\item{ts}{A numeric vector of thresholds}
}
\description{
This function adds the mean, standard deviation (SD) and the
coefficient of variation (CV) to plotting_data
}
\keyword{internal}
