% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{data_imputation}
\alias{data_imputation}
\title{Half-minimum imputation}
\usage{
data_imputation(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{An SE object output from \code{\link[MetAlyzer]{read_metidq()}}.}
}
\value{
An SE object with the imputed `Concentration` column in the aggregated
data accessible via `metalyzer_se@metadata$aggregated_data`.
}
\description{
Impute NA concentrations using half-minimum (HM) and update the
`Concentration` column in the aggregated data in the input `SummarizedExperiment`
(SE) object. If all values of a feature are NA, they stay NA.
}
\keyword{internal}
