% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_draw_FC_network.R
\name{draw_FC_network}
\alias{draw_FC_network}
\title{Draw metabolic pathway}
\usage{
draw_FC_network(
  log2FC_df,
  output,
  q_value = 0.05,
  pathway_file = system.file("extdata", "pathway.xlsx", package = "MetAlyzer"),
  figsize = c(35, 35),
  colbar_width = 2,
  bg_color = "white",
  node_size = 500,
  font_size = 15,
  font_color = "black",
  edge_color = "grey",
  ann_font_size = 25,
  ann_font_color = "blue",
  colors = c("royalblue", "grey80", "tomato2"),
  path_to_conda = NULL
)
}
\arguments{
\item{log2FC_df}{log2FC_df}

\item{output}{output}

\item{q_value}{q_value}

\item{pathway_file}{pathway_file}

\item{figsize}{figsize}

\item{colbar_width}{colbar_width}

\item{bg_color}{bg_color}

\item{node_size}{node_size}

\item{font_size}{font_size}

\item{font_color}{font_color}

\item{edge_color}{edge_color}

\item{ann_font_size}{ann_font_size}

\item{ann_font_color}{ann_font_color}

\item{colors}{colors}

\item{path_to_conda}{path_to_conda}
}
\description{
This function ...
}
\examples{
\dontrun{
print(1)
}
}
