% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer.R
\name{dropInvalidMetabolites}
\alias{dropInvalidMetabolites}
\alias{dropInvalidMetabolites,MetAlyzer-method}
\title{Drop invalid metabolites}
\usage{
dropInvalidMetabolites(
  object,
  aggregated_data,
  filter_col = .data$Valid_Replicates,
  replicate_t = 0
)

\S4method{dropInvalidMetabolites}{MetAlyzer}(
  object,
  aggregated_data,
  filter_col = .data$Valid_Replicates,
  replicate_t = 0
)
}
\arguments{
\item{object}{MetAlyzer object}

\item{aggregated_data}{aggregated_data tibble data frame}

\item{filter_col}{A boolean column from aggregated_data defines the filter}

\item{replicate_t}{A numeric lower threshold between 0 and 1 (t < x) to
determine valid metabolites with a given percentage of valid replicates.}
}
\value{
An updated MetAlyzer object
}
\description{
This method extracts metabolites with a given percentage of valid replicates
from aggregated_data and filters metabolites.
}
\section{Methods (by class)}{
\itemize{
\item \code{dropInvalidMetabolites(MetAlyzer)}: Drop invalid metabolites

}}
\examples{
# To see an example, please check out the vignette.
}
