% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_aggregated_data.R
\name{expand_aggregated_data}
\alias{expand_aggregated_data}
\title{Expand aggregated_data}
\usage{
expand_aggregated_data(metalyzer_se, metadata_col)
}
\arguments{
\item{metalyzer_se}{An SE object output from \code{\link[MetAlyzer]{read_webidq()}}.}

\item{metadata_col}{A vector of character(s) specifying the column in the sample
metadata to add, accessible via `SummarizedExperiment::colData(metalyzer_se)`.}
}
\value{
An SE object with the added sample metadata in the aggregated data accessible
via `metalyzer_se@metadata$aggregated_data`.
}
\description{
Add sample metadata into the aggregated_data in the input `SummarizedExperiment`
(SE) object.
}
\keyword{internal}
