% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{filter_meta_data}
\alias{filter_meta_data}
\title{Filter meta data}
\usage{
filter_meta_data(metalyzer_se, ..., inplace = FALSE)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Use ´col_name´ and condition to filter selected variables.}

\item{inplace}{If FALSE, return a copy. Otherwise, do operation inplace and
return None.}
}
\value{
An updated SummarizedExperiment
}
\description{
This function updates the "Filter" column in meta_data to
filter out samples.
}
\examples{
metalyzer_se <- MetAlyzer::read_metidq(file_path = MetAlyzer::load_demodata_biocrates())

metalyzer_se <- MetAlyzer::filter_meta_data(metalyzer_se, !is.na(Tissue))
metalyzer_se <- MetAlyzer::filter_meta_data(metalyzer_se, `Sample Description` \%in\% 1:6)
}
