% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_metabolites}
\alias{filter_metabolites}
\title{Filter metabolites}
\usage{
filter_metabolites(
  object,
  class_name = "Metabolism Indicators",
  metabo_vec = NULL
)
}
\arguments{
\item{object}{MetAlyzer object}

\item{class_name}{A class to be filtered out}
}
\description{
This function filters out certain classes or metabolites of the metabolites
vector
}
\keyword{internal}
