% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metabolites.R
\name{filter_metabolites}
\alias{filter_metabolites}
\title{Filter metabolites}
\usage{
filter_metabolites(object, class_name, metabo_vec)
}
\arguments{
\item{object}{MetAlyzer object}

\item{class_name}{A class to be filtered out}

\item{metabo_vec}{A character vector defining metabolites to be removed}
}
\description{
This function filters out certain classes or metabolites of the metabolites
vector.
}
\keyword{internal}
