% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetabolites}
\alias{filterMetabolites}
\title{Filter metabolites}
\usage{
filterMetabolites(MetAlyzer, class_name = "Metabolism Indicators")
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{class_name}{A class to be filtered out}
}
\value{
An updated MetAlyzer object
}
\description{
This method calls filter_metabolites() and filters metabolites
}
\examples{
\dontrun{
obj <- filterMetabolites(obj, class_name = "Metabolism Indicators")
}
}
