% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetabolites}
\alias{filterMetabolites}
\title{Filter metabolites}
\usage{
filterMetabolites(
  MetAlyzer,
  class_name = "Metabolism Indicators",
  metabo_vec = NULL
)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{class_name}{A character value defining the class to be removed}

\item{metabo_vec}{A character vector defining metabolites to be removed}
}
\value{
An updated MetAlyzer object
}
\description{
This function filters out certain classes or metabolites of the metabolites
vector
Note: If both "metabo_vec" and "class_name" arguments are used "metabo_vec"
overwrites the "class_name" argument!
}
\examples{
\dontrun{
obj <- filterMetabolites(obj, class_name = "Metabolism Indicators")
obj <- filterMetabolites(obj, metabo_vec = c("C0", "C2", "C3"))
}
}
