% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metabolites-method.R
\name{filterMetabolites}
\alias{filterMetabolites}
\alias{filterMetabolites,MetAlyzer-method}
\title{Filter metabolites}
\usage{
filterMetabolites(
  object,
  class_name = "Metabolism Indicators",
  metabo_vec = NULL
)

\S4method{filterMetabolites}{MetAlyzer}(
  object,
  class_name = "Metabolism Indicators",
  metabo_vec = NULL
)
}
\arguments{
\item{object}{MetAlyzer object}

\item{class_name}{A character value defining the class to be removed}

\item{metabo_vec}{A character vector defining metabolites to be removed}
}
\value{
An updated MetAlyzer object
}
\description{
This method filters out certain classes or metabolites of the metabolites
vector.
Note: "metabo_vec" overwrites the "class_name" argument!
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Filter metabolites
}}

\examples{
\dontrun{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
data <- MetAlyzerDataset(file_path = fpath)

data <- filterMetabolites(data, class_name = "Metabolism Indicators")
# or
data <- filterMetabolites(data, metabo_vec = c("C0", "C2", "C3"))
}
}
