% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetabolites}
\alias{filterMetabolites}
\title{Filter metabolites}
\usage{
filterMetabolites(
  metalyzer,
  drop_metabolites = c("Metabolism Indicators"),
  inplace = TRUE
)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{drop_metabolites}{A character vector defining metabolite classes
or individual metabolites to be removed}
}
\description{
This function filters out certain classes or metabolites
of the metabolites vector. If aggregated_data is not empty,
metabolites and class will also be filtered here.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

drop_metabolites = c("C0", "C2", "C3", "Metabolism Indicators")
filterMetabolites(metalyzer, drop_metabolites)
# or
metalyzer <- filterMetabolites(metalyzer, drop_metabolites, inplace = FALSE)
}
