% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetaData}
\alias{filterMetaData}
\title{Filter meta data}
\usage{
filterMetaData(metalyzer, ..., inplace = TRUE)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{...}{Use ´col_name´ and condition to filter selected variables.}
}
\description{
This function updates the "Filter" column in meta_data to
filter out samples.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

filterMetaData(metalyzer, !is.na(Tissue))
filterMetaData(metalyzer, `Sample Description` \%in\% 1:6)
# or
metalyzer <- filterMetaData(metalyzer, !is.na(Tissue), inplace = FALSE)
metalyzer <- filterMetaData(metalyzer, `Sample Description` \%in\% 1:6, inplace = FALSE)
}
