% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{filterMetaData}
\alias{filterMetaData}
\title{Filter meta data}
\usage{
filterMetaData(MetAlyzer, column, keep = NULL, remove = NULL)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{column}{A column of meta_data for filtering}

\item{keep}{A vector defining which entries to keep from meta_data}

\item{remove}{A vector defining which entries to remove meta_data}
}
\value{
An updated MetAlyzer object
}
\description{
This function updates the "Filter" column in meta_data to filter out samples.
Note: If both "keep" and "remove" arguments are used "keep" overwrites the
"remove" argument.
}
\examples{
\dontrun{
obj <- filterMetaData(obj, column = Methods, keep = 1:6)
# or
obj <- filterMetaData(obj, column = Methods, remove = 7)
}
}
