% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{get_data_range}
\alias{get_data_range}
\title{Get data range}
\usage{
get_data_range(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function extracts rows and column indices to slice the .full_sheet into
metabolites, raw_data and meta_data.
}
\keyword{internal}
