% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_data.R
\name{get_filtered_data}
\alias{get_filtered_data}
\title{Get filtered data}
\usage{
get_filtered_data(object, slot)
}
\arguments{
\item{object}{MetAlyzer object}

\item{slot}{A character value specifying which data frame to slice}
}
\description{
This function returns the filtered meta_data, raw_data or quant_status
data frame.
}
\keyword{internal}
