% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{impute_data}
\alias{impute_data}
\title{Impute aggregated data}
\usage{
impute_data(aggregated_data, impute_to, impute_NA)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}

\item{impute_to}{A numeric value below 1}

\item{impute_NA}{Logical value whether to impute NA values}
}
\description{
This function imputes zero concentration values (Concentration) with the
minimal positive value multiplied by impute_to. If all values are zero or NA,
they are set to NA. The imputed values are added to plotting_data in an extra
column imputed_Conc.
}
\keyword{internal}
