% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_methods.R
\name{imputePlottingData}
\alias{imputePlottingData}
\alias{imputePlottingData,MetAlyzer-method}
\title{Impute plotting data}
\usage{
imputePlottingData(object, ..., i = 0.2, imputeNA = FALSE)

\S4method{imputePlottingData}{MetAlyzer}(object, ..., i = 0.2, imputeNA = FALSE)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{Variables to group by}

\item{i}{A numeric value below 1}

\item{imputeNA}{Logical value whether to impute NA values}
}
\value{
An updated MetAlyzer object
}
\description{
This method imputes zero concentration values (Concentration) with the
minimal positive value multiplied by i. If all values are zero or NA, they
are set to NA. The imputed values are added to plotting_data in an extra
column imp_Conc.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Impute plotting data
}}

\examples{
# To see an example, please check out the vignette.
}
