% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{log2FC}
\alias{log2FC}
\title{Get log2FC Data}
\usage{
log2FC(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}
}
\description{
This function returns the tibble "log2FC".
}
\examples{
metalyzer_se <- MetAlyzer::read_webidq(conc_file_path = MetAlyzer::load_demodata_biocrates())
metalyzer_se@metadata$log2FC <- readRDS(MetAlyzer::toy_diffres())
MetAlyzer::log2FC(metalyzer_se)
}
