% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{metabolites}
\alias{metabolites}
\title{Get metabolites}
\usage{
metabolites(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{

}
\description{
This method returns the filtered metabolites vector
}
