% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{metabolites}
\alias{metabolites}
\alias{metabolites,MetAlyzer-method}
\title{Get metabolites}
\usage{
metabolites(object)

\S4method{metabolites}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The metabolites vector
}
\description{
This method returns the filtered metabolites vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get metabolites
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

metabolites(obj)
}
