% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{metaData}
\alias{metaData}
\title{Get meta data}
\usage{
metaData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
The meta data data frame
}
\description{
This method returns the meta_data data frame with filtered samples (rows) and
all columns
}
\examples{
\dontrun{
metaData(obj)
}
}
