% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{metaData}
\alias{metaData}
\alias{metaData,MetAlyzer-method}
\title{Get meta data}
\usage{
metaData(object)

\S4method{metaData}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The meta_data data frame
}
\description{
This method returns the meta_data data frame with filtered samples (rows) and
all columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get meta data
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

meta_data <- metaData(obj)
head(meta_data)
}
