% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{MetAlyzer_dataset}
\alias{MetAlyzer_dataset}
\title{Open file and read data}
\usage{
MetAlyzer_dataset(...)
}
\arguments{
\item{...}{Declare this function as out of date}
}
\description{
This function was deprecated in version v2.0.0
}
