% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{open_file}
\alias{open_file}
\title{Open Excel file}
\usage{
open_file(starter_list)
}
\arguments{
\item{starter_list}{contains the file path and the sheet index}
}
\description{
This function opens the given MetIDQ output Excel file and reads the full
given sheet.
}
\keyword{internal}
