% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ANOVA.R
\name{perform_ANOVA}
\alias{perform_ANOVA}
\title{ANOVA}
\usage{
perform_ANOVA(object, categorical)
}
\arguments{
\item{object}{MetAlyzer object}

\item{categorical}{A column defining the categorical variable}
}
\description{
This method performs a one-way ANOVA on the grouped plotting_data (the
categorical variable is removed from grouping first). For this, the column
valid_replicates must have at least one entry that is TRUE in each group.
Otherwise, a vector of NA is returned. A Tukey post-hoc test is then used to
determine group names, starting with "A" followed by further letters. These
group names are added to plotting_data in the column ANOVA_group. Thereby,
metabolites can be identified which are significantly higher in one or more
of the categorical variable compared to all other for each metabolite.
}
\keyword{internal}
