% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{performANOVA}
\alias{performANOVA}
\title{ANOVA}
\usage{
performANOVA(MetAlyzer, categorical)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{categorical}{A length-one character vector defining the categorical
variable)}
}
\value{

}
\description{
This method performs a one-way ANOVA adds the column Group to plotting_data
with the results of a Tukey post-hoc test
}
