% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{performANOVA}
\alias{performANOVA}
\title{ANOVA}
\usage{
performANOVA(MetAlyzer, categorical)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{categorical}{A column defining the categorical variable}
}
\value{
An updated MetAlyzer object
}
\description{
This method performs a one-way ANOVA on the grouped plotting_data (the
categorical variable is removed from grouping first). For this, $Valid must
have at least one entry that is TRUE. A Tukey post-hoc test is then used to
determine group names, starting with "A" followed by further letters. These
group names are added to plotting_data in the column Group. Thereby,
metabolites can be identified which are significantly higher in one or more
of the categorical variable compared to all other for each metabolite.
}
\examples{
\dontrun{
obj <- performANOVA(obj, Methods)
}
}
