% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_methods.R
\name{performANOVA}
\alias{performANOVA}
\alias{performANOVA,MetAlyzer-method}
\title{ANOVA}
\usage{
performANOVA(object, categorical)

\S4method{performANOVA}{MetAlyzer}(object, categorical)
}
\arguments{
\item{object}{MetAlyzer object}

\item{categorical}{A column defining the categorical variable}
}
\value{
An updated MetAlyzer object
}
\description{
This method performs a one-way ANOVA on the grouped plotting_data (the
categorical variable is removed from grouping first). For this, the column
valid_replicates must have at least one entry that is TRUE in each group.
Otherwise, a vector of NA is returned. A Tukey post-hoc test is then used to
determine group names, starting with "A" followed by further letters. These
group names are added to plotting_data in the column ANOVA_group. Thereby,
metabolites can be identified which are significantly higher in one or more
of the categorical variable compared to all other for each metabolite.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: ANOVA
}}

\examples{
# To see an example, please check out the vignette.
}
