% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log2FC.R
\name{plot_log2FC}
\alias{plot_log2FC}
\title{Plot log2 fold change}
\usage{
plot_log2FC(
  log2FC_df,
  signif_colors = c(`#5F5F5F` = 1, `#FEBF6E` = 0.1, `#EE5C42` = 0.05, `#8B1A1A` = 0.01),
  hide_labels_for = c(),
  class_colors = "MetAlyzer",
  polarity_file = "MxPQuant500",
  vulcano = FALSE
)
}
\arguments{
\item{log2FC_df}{fold change data frame}

\item{signif_colors}{signif_colors}

\item{hide_labels_for}{vector of Metabolites or Classes for which no labels
are printed}

\item{class_colors}{class_colors}

\item{polarity_file}{polarity_file}
}
\value{
ggplot object
}
\description{
This method plots the log2 fold change for each metabolite.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = extraction_data())
metalyzer_se <- renameMetaData(metalyzer_se, Method = 'Sample Description')

log2FC_df <- calculate_log2FC(metalyzer_se, Method, perc_of_min = 0.2, impute_NA = TRUE)
network <- plot_network(log2FC_df, q_value = 0.05)
}
