% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot Pathway Network}
\usage{
plot_network(
  log2fc_df,
  q_value = 0.05,
  metabolite_text_size = 3,
  connection_width = 0.75,
  pathway_text_size = 6,
  pathway_width = 3,
  plot_column_name = "log2FC",
  exclude_pathways = NULL,
  color_scale = "viridis",
  color_option = "D"
)
}
\arguments{
\item{q_value}{The q-value threshold for significance}

\item{metabolite_text_size}{The text size of metabolite labels}

\item{connection_width}{The line width of connections between metabolites}

\item{pathway_text_size}{The text size of pathway annotations}

\item{pathway_width}{The line width of pathway-specific connection coloring}

\item{plot_column_name}{Column name in the Log2FC dataframe to plot;}

\item{metalyzer_se}{A Metalyzer object}

\item{exlude_pathways}{Pathway names that are exluded from plotting}
}
\value{
ggplot object
}
\description{
This function plots the log2 fold change for each metabolite and visualizes it, in a pathway network.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_mutation_data_xl())
metalyzer_se <- filterMetabolites(
  metalyzer_se,
  drop_metabolites = "Metabolism Indicators"
)
metalyzer_se <- renameMetaData(
  metalyzer_se,
  Mutant_Control = "Sample Description"
)

metalyzer_se <- calculate_log2FC(
  metalyzer_se,
  categorical = "Mutant_Control",
  impute_perc_of_min = 0.2,
  impute_NA = FALSE
)

log2fc_df <- metalyzer_se@metadata$log2FC
network <- plot_network(metalyzer_se, q_value = 0.05)
}
