% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot Pathway Network}
\usage{
plot_network(
  log2fc_df,
  q_value = 0.05,
  metabolite_col_name = "Metabolite",
  values_col_name = "log2FC",
  stat_col_name = "qval",
  metabolite_text_size = 3,
  connection_width = 0.75,
  pathway_text_size = 6,
  pathway_width = 3,
  exclude_pathways = NULL,
  color_scale = "viridis",
  gradient_colors = NULL,
  save_as = NULL,
  folder_name = format(Sys.Date(), "\%Y-\%m-\%d"),
  folder_path = NULL,
  file_name = "network",
  format = "pdf",
  width = 29.7,
  height = 21,
  units = "cm",
  overwrite = FALSE
)
}
\arguments{
\item{log2fc_df}{A dataframe with log2FC, pval, qval, additional columns}

\item{q_value}{The q-value threshold for significance}

\item{metabolite_col_name}{Columnname that holds the Metabolites}

\item{values_col_name}{Column name of a column that holds numeric values, to be plotted \strong{Default = "log2FC"}}

\item{stat_col_name}{Columnname that holds numeric stat values that are used for significance \strong{Default = "qval"}}

\item{metabolite_text_size}{The text size of metabolite labels}

\item{connection_width}{The line width of connections between metabolites}

\item{pathway_text_size}{The text size of pathway annotations}

\item{pathway_width}{The line width of pathway-specific connection coloring}

\item{color_scale}{A string specifying the color scale to use. Options include `"viridis"`, `"plasma"`, `"magma"`, `"inferno"`, `"cividis"`, `"rocket"`, `"mako"`, and `"turbo"`, which use the `viridis` color scales. If `"gradient"` is selected, a custom gradient is applied based on `gradient_colors`.}

\item{gradient_colors}{A vector of length 2 or 3 specifying the colors for a custom gradient. If two colors are provided (`c(low, high)`), `scale_fill_gradient()` is used. If three colors are provided (`c(low, mid, high)`), `scale_fill_gradient2()` is used. If `NULL` or incorrectly specified, the viridis color scale is applied.}

\item{save_as}{\emph{Optional: } Select the file type of output plots. Options are svg, pdf, png or NULL. \strong{Default = "NULL"}}

\item{folder_name}{Name of the folder where the plot will be saved. Special characters will be removed automatically. \strong{Default = date}}

\item{folder_path}{\emph{Optional: } User-defined path where the folder should be created. 
If not provided, results will be saved in `MetAlyzer_results` within the working directory. \strong{Default = NULL}}

\item{file_name}{Name of the output file (without extension). \strong{Default = "network"}}

\item{width}{Width of the saved plot in specified units. \strong{Default = 29.7}}

\item{height}{Height of the saved plot in specified units. \strong{Default = 21.0}}

\item{units}{Units for width and height (e.g., "in", "cm", "mm"). \strong{Default = "cm"}}

\item{overwrite}{Logical: If `TRUE`, overwrite existing files without asking. If `FALSE`, prompt user before overwriting. \strong{Default = FALSE}}

\item{exlude_pathways}{Pathway names that are exluded from plotting}
}
\value{
ggplot object
}
\description{
This function plots the log2 fold change for each metabolite and visualizes it, in a pathway network.
}
\examples{
log2fc_df <- readRDS(load_diffres())
network <- plot_network(log2fc_df, q_value = 0.05)
}
