% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log2FC.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter Plot Visualization}
\usage{
plot_scatter(
  log2fc_df,
  signif_colors = c(`#5F5F5F` = 1, `#FEBF6E` = 0.1, `#EE5C42` = 0.05, `#8B1A1A` = 0.01),
  show_labels_for = NULL
)
}
\arguments{
\item{log2fc_df}{DF with metabolites as row names and columns including log2FC, Class, qval columns.}

\item{signif_colors}{Vector assigning significance values different colors}

\item{show_labels_for}{Vector with Strings of Metabolite names or classes.}
}
\value{
ggplot object
]
}
\description{
This method creates a scatter plot of the log2 fold change for each metabolite.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_mutation_data_xl())
metalyzer_se <- filterMetabolites(
  metalyzer_se,
  drop_metabolites = "Metabolism Indicators"
)
metalyzer_se <- renameMetaData(
  metalyzer_se,
  Mutant_Control = "Sample Description"
)
metalyzer_se <- calculate_log2FC(
  metalyzer_se,
  categorical = "Mutant_Control",
  impute_perc_of_min = 0.2,
  impute_NA = TRUE
)

log2fc_df <- metalyzer_se@metadata$log2FC
scatter <- plot_scatter(log2fc_df)
}
