% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{plot_vulcano}
\alias{plot_vulcano}
\title{Vulcano plot - ggplot}
\usage{
plot_vulcano(
  Log2FCTab,
  x_cutoff = 1.5,
  y_cutoff = 0.05,
  show_labels_for = NULL
)
}
\arguments{
\item{Log2FCTab}{A data frame containing the differential analysis results table
accessible via `metalyzer_se@metadata$log2FC` where `metalyzer_se` is an SE object
output from \code{\link{read_webidq}} and has gone through `calc_log2FC()`.}

\item{x_cutoff, y_cutoff}{Numerical values specifying the cutoffs for log2 fold
changes and q-values. Default is 1.5 and 0.05, respectively.}

\item{show_labels_for}{A vector of characters specifying the names of metabolites
or classes to label. Note that the metabolites of a specified class will be labeled.
Default is NULL.}
}
\value{
A `ggplot` object.
}
\description{
Create a static vulcano plot using `ggplot()`.
}
\keyword{internal}
