% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{plotly_network}
\alias{plotly_network}
\title{Network diagram - ggplotly}
\usage{
plotly_network(
  Log2FCTab,
  q_value = 0.05,
  metabolite_col_name = "Metabolite",
  values_col_name = "log2FC",
  stat_col_name = "qval",
  exclude_pathways = NULL,
  metabolite_node_size = 11,
  connection_width = 1.25,
  pathway_text_size = 20,
  pathway_width = 10,
  plot_height = 800,
  color_scale = "viridis"
)
}
\arguments{
\item{Log2FCTab}{A dataframe with log2FC, qval, additional columns}

\item{q_value}{The q-value threshold for significance}

\item{metabolite_col_name}{Columnname that holds the Metabolites}

\item{values_col_name}{Column name of a column that holds numeric values, to be plotted \strong{Default = "log2FC"}}

\item{stat_col_name}{Columnname that holds numeric stat values that are used for significance \strong{Default = "qval"}}

\item{exclude_pathways}{Pathway names that are exluded from plotting}

\item{metabolite_node_size}{The text size of metabolite nodes}

\item{connection_width}{The line width of connections between metabolites}

\item{pathway_text_size}{The text size of pathway annotations}

\item{pathway_width}{The line width of pathway-specific connection coloring}

\item{plot_height}{The height of the Plot in pixel [px]}

\item{color_scale}{A string specifying the color scale to use. Options include `"viridis"`, `"plasma"`, `"magma"`, `"inferno"`, `"cividis"`, `"rocket"`, `"mako"`, and `"turbo"`, which use the `viridis` color scales.}
}
\description{
This function returns a list with interactive
networkplot based on log2 fold change data.
}
\keyword{internal}
