% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{plotly_scatter}
\alias{plotly_scatter}
\title{Scatter plot - ggplotly}
\usage{
plotly_scatter(Log2FCTab)
}
\arguments{
\item{Log2FCTab}{A data frame containing the differential analysis results table
accessible via `metalyzer_se@metadata$log2FC` where `metalyzer_se` is an SE object
output from \code{\link[MetAlyzer]{read_metidq()}} and has gone through `calc_log2FC()`.}
}
\value{
A list containing a `plotly` (plot) and `ggplot` (legend) object.
}
\description{
Return a list containing an interactive scatter plot made using
`ggplotly()` and its static legend using `ggplot()`. The x-axis represents metabolic
classes, the y-axis shows log2 fold changes, and each point corresponds to a metabolite.
}
\keyword{internal}
