% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_helper.R
\name{plotly_vulcano}
\alias{plotly_vulcano}
\title{Vulcano plot - ggplotly}
\usage{
plotly_vulcano(Log2FCTab, x_cutoff = 1.5, y_cutoff = 0.05)
}
\arguments{
\item{Log2FCTab}{A data frame containing the differential analysis results table
accessible via `metalyzer_se@metadata$log2FC` where `metalyzer_se` is an SE object
output from \code{\link[MetAlyzer]{read_metidq()}} and has gone through `calc_log2FC()`.}

\item{x_cutoff, y_cutoff}{Numerical values specifying the cutoffs for log2 fold
changes and q-values. Default is 1.5 and 0.05, respectively.}
}
\value{
A `plotly` object.
}
\description{
Create an interactive vulcano plot using `ggplotly()`.
}
\keyword{internal}
