% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plotting_data.R
\name{plotting_data}
\alias{plotting_data}
\title{Reshape data}
\usage{
plotting_data(object, ..., ungrouped)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to reshaped data frame}

\item{ungrouped}{A column from meta_data to add to reshaped data frame that
will not be used as grouping variables}
}
\description{
This function reshapes raw_data, quant_status and meta_data and combines them
in a tibble data frame for plotting with 'ggplot2'.
}
\keyword{internal}
