% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{plottingData}
\alias{plottingData}
\title{Get plotting data}
\usage{
plottingData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
The plotting_data data frame
}
\description{
This method returns the plotting_data tibble data frame
}
\examples{
\dontrun{
plottingData(obj)
}
}
