% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{plottingData}
\alias{plottingData}
\alias{plottingData,MetAlyzer-method}
\title{Get plotting data}
\usage{
plottingData(object)

\S4method{plottingData}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The plotting_data data frame
}
\description{
This method returns the plotting_data tibble data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get plotting data
}}

\examples{
\donttest{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)
obj <- createPlottingData(obj, Tissue, Group)

plottingData(obj)
}
}
