% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{quantStatus}
\alias{quantStatus}
\alias{quantStatus,MetAlyzer-method}
\title{Get quantification status}
\usage{
quantStatus(object)

\S4method{quantStatus}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The quant_status data frame
}
\description{
This method returns the quant_status data frame with filtered samples (rows)
and metabolites (columns).
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get quantification status
}}

\examples{
\dontrun{
quantStatus(obj)
}
}
