% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{quantStatus}
\alias{quantStatus}
\title{Get quantification status}
\usage{
quantStatus(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
The quant_status data frame
}
\description{
This method returns the quant_status data frame with filtered samples (rows)
and metabolites (columns)
}
\examples{
\dontrun{
quantStatus(obj)
}
}
