% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{quantStatus}
\alias{quantStatus}
\alias{quantStatus,MetAlyzer-method}
\title{Get quantification status}
\usage{
quantStatus(object)

\S4method{quantStatus}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
The quant_status data frame
}
\description{
This method returns the quant_status data frame with filtered samples (rows)
and metabolites (columns).
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Get quantification status
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

quant_status <- quantStatus(obj)
head(quant_status, c(5, 5))
}
