% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{rawData}
\alias{rawData}
\title{Get raw data}
\usage{
rawData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
The raw_data data frame
}
\description{
This method returns the raw_data data frame with filtered samples (rows) and
metabolites (columns)
}
\examples{
\dontrun{
rawData(obj)
}
}
