% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{read_named_region}
\alias{read_named_region}
\title{Read Named Regions}
\usage{
read_named_region(file_path, named_region)
}
\arguments{
\item{file_path}{The file path of the file}

\item{named_region}{The region name u want to read in}
}
\description{
This function reads in the named regions of an excel file.
}
