% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{read_quant_status}
\alias{read_quant_status}
\title{Read quantification status}
\usage{
read_quant_status(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function gets the background color of each cell in .full_sheet and
assigns it to the corresponding quantification status.
}
\keyword{internal}
