% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{read_quant_status}
\alias{read_quant_status}
\title{Read quantification status}
\usage{
read_quant_status(
  starter_list,
  sheet_dim,
  data_ranges,
  metabolites,
  status_list,
  silent
)
}
\arguments{
\item{starter_list}{starter_list}

\item{sheet_dim}{sheet_dim}

\item{data_ranges}{data_ranges}

\item{metabolites}{metabolites}

\item{status_list}{status_list}

\item{silent}{silent}
}
\description{
This function gets the background color of each cell in 
full_sheet and assigns it to the corresponding quantification status.
}
\keyword{internal}
