% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{readData}
\alias{readData}
\title{Open file and read data}
\usage{
readData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
An updated MetAlyzer object
}
\description{
This method opens the given MetIDQ output Excel file and extracts metabolites,
raw data and meta data
}
\examples{
\dontrun{
obj <- readData(obj)
}
}
