% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{readQuantStatus}
\alias{readQuantStatus}
\title{Read quantification status}
\usage{
readQuantStatus(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{

}
\description{
This function gets the background color of each cell of the opened Excel file and assigns it to the corresponding quantification status
}
