% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{renameMetaData}
\alias{renameMetaData}
\title{Rename meta data}
\usage{
renameMetaData(metalyzer, ..., inplace = TRUE)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{...}{Use new_name = old_name to rename selected variables}
}
\description{
This function renames a column of meta_data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

renameMetaData(metalyzer, Method = `Sample Description`)
# or
metalyzer <- renameMetaData(metalyzer, Method = `Sample Description`, inplace = FALSE)
}
