% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer.R
\name{renameMetaData}
\alias{renameMetaData}
\alias{renameMetaData,MetAlyzer-method}
\title{Rename meta data}
\usage{
renameMetaData(object, ...)

\S4method{renameMetaData}{MetAlyzer}(object, ...)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{Use new_name = old_name to rename selected variables}
}
\value{
An updated MetAlyzer object
}
\description{
This method renames a column of meta_data using rename 'dplyr'.
}
\section{Methods (by class)}{
\itemize{
\item \code{renameMetaData(MetAlyzer)}: Rename meta data

}}
\examples{
\dontrun{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
data <- MetAlyzerDataset(file_path = fpath)

data <- renameMetaData(data, Method = Group)
}
}
