% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{renameMetaData}
\alias{renameMetaData}
\title{Rename meta data}
\usage{
renameMetaData(MetAlyzer, ...)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{...}{Use new_name = old_name to rename selected variables}
}
\value{
An updated MetAlyzer object
}
\description{
This method renames a column of meta_data using rename {dplyr}
}
\examples{
\dontrun{
obj <- renameMetaData(obj, Methods = X1)
}
}
