% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{renameMetaData}
\alias{renameMetaData}
\title{Rename meta data}
\usage{
renameMetaData(metalyzer_se, ..., inplace = FALSE)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Use new_name = old_name to rename selected variables}

\item{inplace}{If FALSE, return a copy. Otherwise, do operation inplace
and return None.}
}
\value{
An updated SummarizedExperiment
}
\description{
This function renames a column of meta_data.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_extraction_data())

metalyzer_se <- renameMetaData(
  metalyzer_se,
  Method = `Sample Description`
)
# or
renameMetaData(metalyzer_se, Model_Organism = Tissue, inplace = TRUE)
}
