% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{resetMetabolites}
\alias{resetMetabolites}
\title{Reset metabolites}
\usage{
resetMetabolites(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
An updated MetAlyzer object
}
\description{
This method resets the filtered metabolites
}
\examples{
\dontrun{
obj <- resetMetabolites(obj)
}
}
