% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites_methods.R
\name{resetMetabolites}
\alias{resetMetabolites}
\alias{resetMetabolites,MetAlyzer-method}
\title{Reset metabolites}
\usage{
resetMetabolites(object)

\S4method{resetMetabolites}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
An updated MetAlyzer object
}
\description{
This method resets the filtering of metabolites.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Reset metabolites
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

obj <- resetMetabolites(obj)
}
