% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{resetMetaData}
\alias{resetMetaData}
\title{Reset meta data}
\usage{
resetMetaData(metalyzer, inplace = TRUE)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\description{
This function resets the filter of meta_data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

resetMetaData(metalyzer)
# or
metalyzer <- resetMetaData(metalyzer, inplace = FALSE)
}
