% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_methods.R
\name{resetMetaData}
\alias{resetMetaData}
\alias{resetMetaData,MetAlyzer-method}
\title{Reset meta data}
\usage{
resetMetaData(object)

\S4method{resetMetaData}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
An updated MetAlyzer object
}
\description{
This method resets the filter of meta_data.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Reset meta data
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

obj <- resetMetaData(obj)
}
