% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshape data}
\usage{
reshape_data(object, ..., id_col)
}
\arguments{
\item{object}{MetAlyzer object}

\item{...}{A selection of columns from meta_data to add to aggregated data frame}

\item{id_col}{Sample ID column from meta_data that is added to aggregated data
frame but is not used as grouping variable.}
}
\description{
This function reshapes raw_data, quant_status and meta_data and combines them
in a tibble data frame for plotting with 'ggplot2'.
}
\keyword{internal}
