% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cv.R
\name{set_threshold}
\alias{set_threshold}
\title{Threshold CV}
\usage{
set_threshold(x, cv_threshs)
}
\arguments{
\item{x}{A CV value}

\item{cv_threshs}{A numeric vector of thresholds}
}
\description{
This function assigns a CV value according to a vector of thresholds.
}
\keyword{internal}
