% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_methods.R
\name{setPlottingData}
\alias{setPlottingData}
\alias{setPlottingData,MetAlyzer-method}
\title{Update plotting data}
\usage{
setPlottingData(object, plotting_data)

\S4method{setPlottingData}{MetAlyzer}(object, plotting_data)
}
\arguments{
\item{object}{MetAlyzer object}

\item{plotting_data}{An updated plotting_data tibble data frame}
}
\value{
An updated MetAlyzer object
}
\description{
This method replaces plotting_data with an updated version.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Update plotting data
}}

\examples{
# To see an example, please check out the vignette.
}
