% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{show,MetAlyzer-method}
\alias{show,MetAlyzer-method}
\title{Show a MetAlyzer object}
\usage{
\S4method{show}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
A summary of the MetAlyzer object
}
\description{
This method shows a summary of MetAlyzer slot values
}
\examples{
\dontrun{
show(obj)
# or
obj
}
}
