% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{show,MetAlyzer-method}
\alias{show,MetAlyzer-method}
\title{Show a 'MetAlyzer' object}
\usage{
\S4method{show}{MetAlyzer}(object)
}
\arguments{
\item{metalyzer}{MetAlyzer object}
}
\value{
A summary of the MetAlyzer object
}
\description{
This function shows a summary of the 'MetAlyzer' slot values.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

show(metalyzer)
# or
metalyzer
}
