% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{show,MetAlyzer-method}
\alias{show,MetAlyzer-method}
\title{Show a 'MetAlyzer' object}
\usage{
\S4method{show}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
A summary of the MetAlyzer object
}
\description{
This method shows a summary of 'MetAlyzer' slot values.
}
\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

show(obj)
# or
obj
}
