% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_obj}
\alias{show_obj}
\title{Show the 'MetAlyzer' object}
\usage{
show_obj(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function shows a summary of 'MetAlyzer' slot values.
}
\keyword{internal}
