% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_dataset.R
\name{slice_meta_data}
\alias{slice_meta_data}
\title{Slice meta data}
\usage{
slice_meta_data(full_sheet, data_ranges)
}
\arguments{
\item{full_sheet}{full_sheet}

\item{data_ranges}{data_ranges}
}
\description{
This function slices meta data from .full_sheet into meta_data.
}
\keyword{internal}
