% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{slice_metabolties}
\alias{slice_metabolties}
\title{Slice metabolites}
\usage{
slice_metabolties(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function extracts metabolites with their corresponding metabolite class
from .full_sheet into metabolites.
}
\keyword{internal}
