% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalyzer_dataset.R
\name{slice_raw_data}
\alias{slice_raw_data}
\title{Slice raw data}
\usage{
slice_raw_data(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\description{
This function slices measurements from .full_sheet into raw_data.
}
\keyword{internal}
