% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{summariseQuantData}
\alias{summariseQuantData}
\title{Summarize quantification status}
\usage{
summariseQuantData(MetAlyzer)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}
}
\value{
A summary of the quantification status
}
\description{
This method lists the number of each quantification status and its percentage
}
\examples{
\dontrun{
summariseQuantData(obj)
}
}
