% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{summariseQuantData}
\alias{summariseQuantData}
\alias{summariseQuantData,MetAlyzer-method}
\title{Summarize quantification status}
\usage{
summariseQuantData(object)

\S4method{summariseQuantData}{MetAlyzer}(object)
}
\arguments{
\item{object}{MetAlyzer object}
}
\value{
A summary of the quantification status
}
\description{
This method lists the number of each quantification status and its
percentage.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Summarize quantification status
}}

\examples{
fpath <- system.file("extdata", "example_data.xlsx", package = "MetAlyzer")
obj <- MetAlyzerDataset(file_path = fpath)

summariseQuantData(obj)
}
