% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{summarizeQuantData}
\alias{summarizeQuantData}
\title{Summarize quantification status}
\usage{
summarizeQuantData(metalyzer_se)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}
}
\description{
This function lists the number of each quantification status and
its percentage.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_extraction_data())

summarizeQuantData(metalyzer_se)
}
