% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_fpaths.R
\name{toy_diffres}
\alias{toy_diffres}
\title{Get example log2fc data}
\usage{
toy_diffres()
}
\value{
toy_diffres.rds file path
}
\description{
This function returns the log2fc dataframe of the Metalyzer_demo dataset_biocrates MxP Quant 500 XL_2025-04 file, created with the MetaVizPro package.
}
\examples{
fpath <- MetAlyzer::toy_diffres()
}
