% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{transform}
\alias{transform}
\title{Transformation}
\usage{
transform(vec, func)
}
\arguments{
\item{vec}{MetAlyzer metalyzer}

\item{func}{A function for transformation}
}
\description{
This function performs transformation of imputed concentration values
(imputed_Conc).
}
\keyword{internal}
