% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_plotting_data.R
\name{transform_plotting_data}
\alias{transform_plotting_data}
\title{Transform plotting data}
\usage{
transform_plotting_data(object, func)
}
\arguments{
\item{object}{MetAlyzer object}

\item{func}{A function to transform concentration values with}
}
\description{
This method performs a transformation of imputed concentration values
(imp_Conc) with a given function. NA values are skipped. The transformed
values are added to plotting_data in an extra column transf_Conc.
}
\keyword{internal}
