% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{transformPlottingData}
\alias{transformPlottingData}
\title{log transform plotting data}
\usage{
transformPlottingData(MetAlyzer, func = log2)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{func}{A logarithmic function to transform concentration values with
number of samples}
}
\value{

}
\description{
This method adds the column transf_Conc to plotting_data with the logarithmic
transformation of imputed concentration values (imp_Conc)
}
