% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_methods.R
\name{transformPlottingData}
\alias{transformPlottingData}
\alias{transformPlottingData,MetAlyzer-method}
\title{Transform plotting data}
\usage{
transformPlottingData(object, func = log2)

\S4method{transformPlottingData}{MetAlyzer}(object, func = log2)
}
\arguments{
\item{object}{MetAlyzer object}

\item{func}{A function to transform concentration values with
number of samples}
}
\value{
An updated MetAlyzer object
}
\description{
This method performs a transformation of imputed concentration values
(imp_Conc) with a given funciton. NA values are skipped. The transformed
values are added to plotting_data in an extra column transf_Conc.
}
\section{Methods (by class)}{
\itemize{
\item \code{MetAlyzer}: Transform plotting data
}}

\examples{
# To see an example, please check out the vignette.
}
