% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{update_meta_data}
\alias{update_meta_data}
\title{Update meta data}
\usage{
update_meta_data(metalyzer_se, ..., inplace = FALSE)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Use ´new_col_name = new_column´ to rename selected variables}

\item{inplace}{If FALSE, return a copy. Otherwise, do operation inplace
and return None.}
}
\value{
An updated SummarizedExperiment
}
\description{
This function adds another column to filtered meta_data.
}
\examples{
metalyzer_se <- MetAlyzer::read_metidq(file_path = MetAlyzer::load_demodata_biocrates())

metalyzer_se <- MetAlyzer::update_meta_data(
  metalyzer_se,
  Date = Sys.Date(), Analyzed = TRUE
)
}
