% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{updateMetaData}
\alias{updateMetaData}
\title{Update meta data}
\usage{
updateMetaData(metalyzer, ..., inplace = TRUE)
}
\arguments{
\item{metalyzer}{MetAlyzer object}

\item{...}{Use ´new_col_name = new_column´ to rename selected variables}
}
\description{
This function adds another column to filtered meta_data.
}
\examples{
metalyzer <- MetAlyzer_dataset(file_path = extraction_data())

updateMetaData(metalyzer, Date = Sys.Date(), Analyzed = TRUE)
# or
metalyzer <- updateMetaData(metalyzer, Date = Sys.Date(), Analyzed = TRUE, inplace = FALSE)
}
