% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer_handler.R
\name{updateMetaData}
\alias{updateMetaData}
\title{Update meta data}
\usage{
updateMetaData(metalyzer_se, ..., inplace = FALSE)
}
\arguments{
\item{metalyzer_se}{SummarizedExperiment}

\item{...}{Use ´new_col_name = new_column´ to rename selected variables}

\item{inplace}{If FALSE, return a copy. Otherwise, do operation inplace
and return None.}
}
\value{
An updated SummarizedExperiment
}
\description{
This function adds another column to filtered meta_data.
}
\examples{
metalyzer_se <- MetAlyzer_dataset(file_path = example_extraction_data())

metalyzer_se <- updateMetaData(
  metalyzer_se,
  Date = Sys.Date(), Analyzed = TRUE
)
# or
updateMetaData(
  metalyzer_se,
  Date = Sys.Date(), Analyzed = TRUE, inplace = TRUE
)
}
