% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetAlyzer-class.R
\name{updatePlottingData}
\alias{updatePlottingData}
\title{Update plotting data}
\usage{
updatePlottingData(MetAlyzer, name, new_colum)
}
\arguments{
\item{MetAlyzer}{MetAlyzer object}

\item{name}{A length-one character vector giving the new column name}

\item{new_colum}{A vector for the new column (length has to be same as the
number of samples)}
}
\value{

}
\description{
This method adds or modifies a column to/of plotting_data
}
