% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{valid_measurement}
\alias{valid_measurement}
\title{Add valid filter}
\usage{
valid_measurement(aggregated_data, valid_vec, ko_vec, valid_thresh)
}
\arguments{
\item{aggregated_data}{aggregated_data tibble data frame}

\item{valid_vec}{A character vector containing each quantification status that
is considered to be a valid measurement}

\item{ko_vec}{A character vector containing each quantification status that
invalidates a set of repetitions once it is encountered.}

\item{valid_thresh}{A numeric threshold}
}
\description{
This function adds a filter column based on the quantification status. The
filter is True if the percentage of measurements with a quantification status
part of valid_vec is greater than the threshold valid_thresh.
}
\keyword{internal}
