% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plotting_data.R
\name{valid_measurement}
\alias{valid_measurement}
\title{Add valid filter}
\usage{
valid_measurement(plotting_data, valid_vec, t)
}
\arguments{
\item{plotting_data}{plotting_data tibble data frame}

\item{valid_vec}{A character vector containing each quantification status that
is considered to be a valid measurement}

\item{t}{A numeric threshold}
}
\description{
This function adds a filter column based on the quantification status. The
filter is True if the percentage of measurements with a quantification status
part of valid_vec is greater than the threshold t.
}
\keyword{internal}
