% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{zero_imputation}
\alias{zero_imputation}
\title{Zero imputation}
\usage{
zero_imputation(vec, impute_to, impute_NA)
}
\arguments{
\item{vec}{A numeric vector containing the concentration values}

\item{impute_to}{A numeric value below 1}

\item{impute_NA}{Logical value whether to impute NA values}
}
\description{
This function performs zero imputation with the minimal positive value times
impute_to.
}
\keyword{internal}
